move SubQueryResult.cpp to SubSearchResult.cpp
