+++ internal/core/src/query/SubSearchResult.cpp
@@ -10,7 +10,7 @@
 // or implied. See the License for the specific language governing permissions and limitations under the License
 
 #include "exceptions/EasyAssert.h"
-#include "query/SubQueryResult.h"
+#include "query/SubSearchResult.h"
 #include "segcore/Reduce.h"
 
 namespace milvus::query {
