+++ internal/core/src/query/SubSearchResult.h
